Restructure API documentation, rename training->running
