+++ docs/source/api-docs/data.rst
@@ -1,0 +1,39 @@
+data
+============
+
+.. automodule:: data
+    :members:
+    :undoc-members:
+    :show-inheritance:
+
+data\_layer
+-----------------------
+
+.. automodule:: data.data_layer
+    :members:
+    :undoc-members:
+    :show-inheritance:
+
+speech2text
+-----------------------
+
+.. automodule:: data.speech2text
+    :members:
+    :undoc-members:
+    :show-inheritance:
+
+speech\_utils
+-------------------------
+
+.. automodule:: data.speech_utils
+    :members:
+    :undoc-members:
+    :show-inheritance:
+
+text2text
+---------------------
+
+.. automodule:: data.text2text
+    :members:
+    :undoc-members:
+    :show-inheritance:
