+++ docs/source/api-docs/decoders.rst
@@ -1,0 +1,39 @@
+decoders
+========
+
+.. automodule:: decoders
+    :members:
+    :undoc-members:
+    :show-inheritance:
+
+decoder
+-----------------------
+
+.. automodule:: decoders.decoder
+    :members:
+    :undoc-members:
+    :show-inheritance:
+
+fc\_decoder
+---------------------------
+
+.. automodule:: decoders.fc_decoder
+    :members:
+    :undoc-members:
+    :show-inheritance:
+
+rnn\_decoders
+-----------------------------
+
+.. automodule:: decoders.rnn_decoders
+    :members:
+    :undoc-members:
+    :show-inheritance:
+
+transformer\_decoders
+-------------------------------------
+
+.. automodule:: decoders.transformer_decoders
+    :members:
+    :undoc-members:
+    :show-inheritance:
