+++ docs/source/api-docs/encoders.rst
@@ -1,0 +1,39 @@
+encoders
+========
+
+.. automodule:: encoders
+    :members:
+    :undoc-members:
+    :show-inheritance:
+
+encoder
+-----------------------
+
+.. automodule:: encoders.encoder
+    :members:
+    :undoc-members:
+    :show-inheritance:
+
+ds2\_encoder
+----------------------------
+
+.. automodule:: encoders.ds2_encoder
+    :members:
+    :undoc-members:
+    :show-inheritance:
+
+rnn\_encoders
+-----------------------------
+
+.. automodule:: encoders.rnn_encoders
+    :members:
+    :undoc-members:
+    :show-inheritance:
+
+transformer\_encoders
+-------------------------------------
+
+.. automodule:: encoders.transformer_encoders
+    :members:
+    :undoc-members:
+    :show-inheritance:
