+++ docs/source/api-docs/losses.rst
@@ -1,0 +1,31 @@
+losses
+======
+
+.. automodule:: losses
+    :members:
+    :undoc-members:
+    :show-inheritance:
+
+loss
+------------------
+
+.. automodule:: losses.loss
+    :members:
+    :undoc-members:
+    :show-inheritance:
+
+ctc\_loss
+-----------------------
+
+.. automodule:: losses.ctc_loss
+    :members:
+    :undoc-members:
+    :show-inheritance:
+
+sequence\_loss
+----------------------------
+
+.. automodule:: losses.sequence_loss
+    :members:
+    :undoc-members:
+    :show-inheritance:
