+++ docs/source/api-docs/models.rst
@@ -1,0 +1,39 @@
+models
+======
+
+.. automodule:: models
+    :members:
+    :undoc-members:
+    :show-inheritance:
+
+model
+-------------------
+
+.. automodule:: models.model
+    :members:
+    :undoc-members:
+    :show-inheritance:
+
+seq2seq
+---------------------
+
+.. automodule:: models.seq2seq
+    :members:
+    :undoc-members:
+    :show-inheritance:
+
+speech2text
+-------------------------
+
+.. automodule:: models.speech2text
+    :members:
+    :undoc-members:
+    :show-inheritance:
+
+text2text
+-----------------------
+
+.. automodule:: models.text2text
+    :members:
+    :undoc-members:
+    :show-inheritance:
