--- docs/source/api-docs/open_seq2seq.data.rst
@@ -1,70 +1,0 @@
-open\_seq2seq.data package
-==========================
-
-Submodules
-----------
-
-open\_seq2seq.data.data\_layer module
--------------------------------------
-
-.. automodule:: open_seq2seq.data.data_layer
-    :members:
-    :undoc-members:
-    :show-inheritance:
-
-open\_seq2seq.data.speech2text module
--------------------------------------
-
-.. automodule:: open_seq2seq.data.speech2text
-    :members:
-    :undoc-members:
-    :show-inheritance:
-
-open\_seq2seq.data.speech2text\_test module
--------------------------------------------
-
-.. automodule:: open_seq2seq.data.speech2text_test
-    :members:
-    :undoc-members:
-    :show-inheritance:
-
-open\_seq2seq.data.speech\_utils module
----------------------------------------
-
-.. automodule:: open_seq2seq.data.speech_utils
-    :members:
-    :undoc-members:
-    :show-inheritance:
-
-open\_seq2seq.data.speech\_utils\_test module
----------------------------------------------
-
-.. automodule:: open_seq2seq.data.speech_utils_test
-    :members:
-    :undoc-members:
-    :show-inheritance:
-
-open\_seq2seq.data.text2text module
------------------------------------
-
-.. automodule:: open_seq2seq.data.text2text
-    :members:
-    :undoc-members:
-    :show-inheritance:
-
-open\_seq2seq.data.text2text\_test module
------------------------------------------
-
-.. automodule:: open_seq2seq.data.text2text_test
-    :members:
-    :undoc-members:
-    :show-inheritance:
-
-
-Module contents
----------------
-
-.. automodule:: open_seq2seq.data
-    :members:
-    :undoc-members:
-    :show-inheritance:
