--- docs/source/api-docs/open_seq2seq.decoders.rst
@@ -1,54 +1,0 @@
-open\_seq2seq.decoders package
-==============================
-
-Submodules
-----------
-
-open\_seq2seq.decoders.decoder module
--------------------------------------
-
-.. automodule:: open_seq2seq.decoders.decoder
-    :members:
-    :undoc-members:
-    :show-inheritance:
-
-open\_seq2seq.decoders.fc\_decoder module
------------------------------------------
-
-.. automodule:: open_seq2seq.decoders.fc_decoder
-    :members:
-    :undoc-members:
-    :show-inheritance:
-
-open\_seq2seq.decoders.rnn\_decoders module
--------------------------------------------
-
-.. automodule:: open_seq2seq.decoders.rnn_decoders
-    :members:
-    :undoc-members:
-    :show-inheritance:
-
-open\_seq2seq.decoders.transformer\_decoders module
----------------------------------------------------
-
-.. automodule:: open_seq2seq.decoders.transformer_decoders
-    :members:
-    :undoc-members:
-    :show-inheritance:
-
-open\_seq2seq.decoders.transformer\_decoders\_test module
----------------------------------------------------------
-
-.. automodule:: open_seq2seq.decoders.transformer_decoders_test
-    :members:
-    :undoc-members:
-    :show-inheritance:
-
-
-Module contents
----------------
-
-.. automodule:: open_seq2seq.decoders
-    :members:
-    :undoc-members:
-    :show-inheritance:
