--- docs/source/api-docs/open_seq2seq.encoders.rst
@@ -1,46 +1,0 @@
-open\_seq2seq.encoders package
-==============================
-
-Submodules
-----------
-
-open\_seq2seq.encoders.ds2\_encoder module
-------------------------------------------
-
-.. automodule:: open_seq2seq.encoders.ds2_encoder
-    :members:
-    :undoc-members:
-    :show-inheritance:
-
-open\_seq2seq.encoders.encoder module
--------------------------------------
-
-.. automodule:: open_seq2seq.encoders.encoder
-    :members:
-    :undoc-members:
-    :show-inheritance:
-
-open\_seq2seq.encoders.rnn\_encoders module
--------------------------------------------
-
-.. automodule:: open_seq2seq.encoders.rnn_encoders
-    :members:
-    :undoc-members:
-    :show-inheritance:
-
-open\_seq2seq.encoders.transformer\_encoders module
----------------------------------------------------
-
-.. automodule:: open_seq2seq.encoders.transformer_encoders
-    :members:
-    :undoc-members:
-    :show-inheritance:
-
-
-Module contents
----------------
-
-.. automodule:: open_seq2seq.encoders
-    :members:
-    :undoc-members:
-    :show-inheritance:
