--- docs/source/api-docs/open_seq2seq.losses.rst
@@ -1,46 +1,0 @@
-open\_seq2seq.losses package
-============================
-
-Submodules
-----------
-
-open\_seq2seq.losses.ctc\_loss module
--------------------------------------
-
-.. automodule:: open_seq2seq.losses.ctc_loss
-    :members:
-    :undoc-members:
-    :show-inheritance:
-
-open\_seq2seq.losses.loss module
---------------------------------
-
-.. automodule:: open_seq2seq.losses.loss
-    :members:
-    :undoc-members:
-    :show-inheritance:
-
-open\_seq2seq.losses.sequence\_loss module
-------------------------------------------
-
-.. automodule:: open_seq2seq.losses.sequence_loss
-    :members:
-    :undoc-members:
-    :show-inheritance:
-
-open\_seq2seq.losses.sequence\_loss\_test module
-------------------------------------------------
-
-.. automodule:: open_seq2seq.losses.sequence_loss_test
-    :members:
-    :undoc-members:
-    :show-inheritance:
-
-
-Module contents
----------------
-
-.. automodule:: open_seq2seq.losses
-    :members:
-    :undoc-members:
-    :show-inheritance:
