--- docs/source/api-docs/open_seq2seq.models.rst
@@ -1,62 +1,0 @@
-open\_seq2seq.models package
-============================
-
-Submodules
-----------
-
-open\_seq2seq.models.model module
----------------------------------
-
-.. automodule:: open_seq2seq.models.model
-    :members:
-    :undoc-members:
-    :show-inheritance:
-
-open\_seq2seq.models.seq2seq module
------------------------------------
-
-.. automodule:: open_seq2seq.models.seq2seq
-    :members:
-    :undoc-members:
-    :show-inheritance:
-
-open\_seq2seq.models.speech2text module
----------------------------------------
-
-.. automodule:: open_seq2seq.models.speech2text
-    :members:
-    :undoc-members:
-    :show-inheritance:
-
-open\_seq2seq.models.speech2text\_test module
----------------------------------------------
-
-.. automodule:: open_seq2seq.models.speech2text_test
-    :members:
-    :undoc-members:
-    :show-inheritance:
-
-open\_seq2seq.models.text2text module
--------------------------------------
-
-.. automodule:: open_seq2seq.models.text2text
-    :members:
-    :undoc-members:
-    :show-inheritance:
-
-open\_seq2seq.models.text2text\_test module
--------------------------------------------
-
-.. automodule:: open_seq2seq.models.text2text_test
-    :members:
-    :undoc-members:
-    :show-inheritance:
-
-
-Module contents
----------------
-
-.. automodule:: open_seq2seq.models
-    :members:
-    :undoc-members:
-    :show-inheritance:
