--- docs/source/api-docs/open_seq2seq.optimizers.rst
@@ -1,54 +1,0 @@
-open\_seq2seq.optimizers package
-================================
-
-Submodules
-----------
-
-open\_seq2seq.optimizers.automatic\_loss\_scaler module
--------------------------------------------------------
-
-.. automodule:: open_seq2seq.optimizers.automatic_loss_scaler
-    :members:
-    :undoc-members:
-    :show-inheritance:
-
-open\_seq2seq.optimizers.lr\_policies module
---------------------------------------------
-
-.. automodule:: open_seq2seq.optimizers.lr_policies
-    :members:
-    :undoc-members:
-    :show-inheritance:
-
-open\_seq2seq.optimizers.mp\_wrapper module
--------------------------------------------
-
-.. automodule:: open_seq2seq.optimizers.mp_wrapper
-    :members:
-    :undoc-members:
-    :show-inheritance:
-
-open\_seq2seq.optimizers.mp\_wrapper\_test module
--------------------------------------------------
-
-.. automodule:: open_seq2seq.optimizers.mp_wrapper_test
-    :members:
-    :undoc-members:
-    :show-inheritance:
-
-open\_seq2seq.optimizers.optimizers module
-------------------------------------------
-
-.. automodule:: open_seq2seq.optimizers.optimizers
-    :members:
-    :undoc-members:
-    :show-inheritance:
-
-
-Module contents
----------------
-
-.. automodule:: open_seq2seq.optimizers
-    :members:
-    :undoc-members:
-    :show-inheritance:
