--- docs/source/api-docs/open_seq2seq.parts.experimental_rnn_cells.rst
@@ -1,46 +1,0 @@
-open\_seq2seq.parts.experimental\_rnn\_cells package
-====================================================
-
-Submodules
-----------
-
-open\_seq2seq.parts.experimental\_rnn\_cells.flstm module
----------------------------------------------------------
-
-.. automodule:: open_seq2seq.parts.experimental_rnn_cells.flstm
-    :members:
-    :undoc-members:
-    :show-inheritance:
-
-open\_seq2seq.parts.experimental\_rnn\_cells.glstm module
----------------------------------------------------------
-
-.. automodule:: open_seq2seq.parts.experimental_rnn_cells.glstm
-    :members:
-    :undoc-members:
-    :show-inheritance:
-
-open\_seq2seq.parts.experimental\_rnn\_cells.gnmt module
---------------------------------------------------------
-
-.. automodule:: open_seq2seq.parts.experimental_rnn_cells.gnmt
-    :members:
-    :undoc-members:
-    :show-inheritance:
-
-open\_seq2seq.parts.experimental\_rnn\_cells.slstm module
----------------------------------------------------------
-
-.. automodule:: open_seq2seq.parts.experimental_rnn_cells.slstm
-    :members:
-    :undoc-members:
-    :show-inheritance:
-
-
-Module contents
----------------
-
-.. automodule:: open_seq2seq.parts.experimental_rnn_cells
-    :members:
-    :undoc-members:
-    :show-inheritance:
