--- docs/source/api-docs/open_seq2seq.parts.rst
@@ -1,77 +1,0 @@
-open\_seq2seq.parts package
-===========================
-
-Subpackages
------------
-
-.. toctree::
-
-    open_seq2seq.parts.experimental_rnn_cells
-
-Submodules
-----------
-
-open\_seq2seq.parts.attention module
-------------------------------------
-
-.. automodule:: open_seq2seq.parts.attention
-    :members:
-    :undoc-members:
-    :show-inheritance:
-
-open\_seq2seq.parts.attention\_test module
-------------------------------------------
-
-.. automodule:: open_seq2seq.parts.attention_test
-    :members:
-    :undoc-members:
-    :show-inheritance:
-
-open\_seq2seq.parts.attention\_wrapper module
----------------------------------------------
-
-.. automodule:: open_seq2seq.parts.attention_wrapper
-    :members:
-    :undoc-members:
-    :show-inheritance:
-
-open\_seq2seq.parts.common module
----------------------------------
-
-.. automodule:: open_seq2seq.parts.common
-    :members:
-    :undoc-members:
-    :show-inheritance:
-
-open\_seq2seq.parts.common\_test module
----------------------------------------
-
-.. automodule:: open_seq2seq.parts.common_test
-    :members:
-    :undoc-members:
-    :show-inheritance:
-
-open\_seq2seq.parts.t2t\_timing\_signal module
-----------------------------------------------
-
-.. automodule:: open_seq2seq.parts.t2t_timing_signal
-    :members:
-    :undoc-members:
-    :show-inheritance:
-
-open\_seq2seq.parts.utils module
---------------------------------
-
-.. automodule:: open_seq2seq.parts.utils
-    :members:
-    :undoc-members:
-    :show-inheritance:
-
-
-Module contents
----------------
-
-.. automodule:: open_seq2seq.parts
-    :members:
-    :undoc-members:
-    :show-inheritance:
