--- docs/source/api-docs/open_seq2seq.rst
@@ -1,37 +1,0 @@
-open\_seq2seq package
-=====================
-
-Subpackages
------------
-
-.. toctree::
-
-    open_seq2seq.data
-    open_seq2seq.decoders
-    open_seq2seq.encoders
-    open_seq2seq.losses
-    open_seq2seq.models
-    open_seq2seq.optimizers
-    open_seq2seq.parts
-    open_seq2seq.test_utils
-    open_seq2seq.training
-
-Submodules
-----------
-
-open\_seq2seq.utils module
---------------------------
-
-.. automodule:: open_seq2seq.utils
-    :members:
-    :undoc-members:
-    :show-inheritance:
-
-
-Module contents
----------------
-
-.. automodule:: open_seq2seq
-    :members:
-    :undoc-members:
-    :show-inheritance:
