--- docs/source/api-docs/open_seq2seq.test_utils.rst
@@ -1,30 +1,0 @@
-open\_seq2seq.test\_utils package
-=================================
-
-Submodules
-----------
-
-open\_seq2seq.test\_utils.create\_reversed\_examples module
------------------------------------------------------------
-
-.. automodule:: open_seq2seq.test_utils.create_reversed_examples
-    :members:
-    :undoc-members:
-    :show-inheritance:
-
-open\_seq2seq.test\_utils.test\_speech\_config module
------------------------------------------------------
-
-.. automodule:: open_seq2seq.test_utils.test_speech_config
-    :members:
-    :undoc-members:
-    :show-inheritance:
-
-
-Module contents
----------------
-
-.. automodule:: open_seq2seq.test_utils
-    :members:
-    :undoc-members:
-    :show-inheritance:
