--- docs/source/api-docs/open_seq2seq.training.rst
@@ -1,38 +1,0 @@
-open\_seq2seq.training package
-==============================
-
-Submodules
-----------
-
-open\_seq2seq.training.funcs module
------------------------------------
-
-.. automodule:: open_seq2seq.training.funcs
-    :members:
-    :undoc-members:
-    :show-inheritance:
-
-open\_seq2seq.training.model\_builders module
----------------------------------------------
-
-.. automodule:: open_seq2seq.training.model_builders
-    :members:
-    :undoc-members:
-    :show-inheritance:
-
-open\_seq2seq.training.utils module
------------------------------------
-
-.. automodule:: open_seq2seq.training.utils
-    :members:
-    :undoc-members:
-    :show-inheritance:
-
-
-Module contents
----------------
-
-.. automodule:: open_seq2seq.training
-    :members:
-    :undoc-members:
-    :show-inheritance:
