+++ docs/source/api-docs/optimizers.rst
@@ -1,0 +1,40 @@
+optimizers
+==========
+
+.. automodule:: optimizers
+    :members:
+    :undoc-members:
+    :show-inheritance:
+
+
+optimizers
+----------------------------
+
+.. automodule:: optimizers.optimizers
+    :members:
+    :undoc-members:
+    :show-inheritance:
+
+mp\_wrapper
+-----------------------------
+
+.. automodule:: optimizers.mp_wrapper
+    :members:
+    :undoc-members:
+    :show-inheritance:
+
+automatic\_loss\_scaler
+-----------------------------------------
+
+.. automodule:: optimizers.automatic_loss_scaler
+    :members:
+    :undoc-members:
+    :show-inheritance:
+
+lr\_policies
+------------------------------
+
+.. automodule:: optimizers.lr_policies
+    :members:
+    :undoc-members:
+    :show-inheritance:
