+++ docs/source/api-docs/parts.experimental_rnn_cells.rst
@@ -1,0 +1,39 @@
+experimental\_rnn\_cells
+======================================
+
+.. automodule:: parts.experimental_rnn_cells
+    :members:
+    :undoc-members:
+    :show-inheritance:
+
+flstm
+-------------------------------------------
+
+.. automodule:: parts.experimental_rnn_cells.flstm
+    :members:
+    :undoc-members:
+    :show-inheritance:
+
+glstm
+-------------------------------------------
+
+.. automodule:: parts.experimental_rnn_cells.glstm
+    :members:
+    :undoc-members:
+    :show-inheritance:
+
+gnmt
+------------------------------------------
+
+.. automodule:: parts.experimental_rnn_cells.gnmt
+    :members:
+    :undoc-members:
+    :show-inheritance:
+
+slstm
+-------------------------------------------
+
+.. automodule:: parts.experimental_rnn_cells.slstm
+    :members:
+    :undoc-members:
+    :show-inheritance:
