+++ docs/source/api-docs/parts.rst
@@ -1,0 +1,51 @@
+parts
+=====
+
+.. automodule:: parts
+    :members:
+    :undoc-members:
+    :show-inheritance:
+
+.. toctree::
+
+    parts.experimental_rnn_cells
+
+attention
+----------------------
+
+.. automodule:: parts.attention
+    :members:
+    :undoc-members:
+    :show-inheritance:
+
+attention\_wrapper
+-------------------------------
+
+.. automodule:: parts.attention_wrapper
+    :members:
+    :undoc-members:
+    :show-inheritance:
+
+common
+-------------------
+
+.. automodule:: parts.common
+    :members:
+    :undoc-members:
+    :show-inheritance:
+
+t2t\_timing\_signal
+--------------------------------
+
+.. automodule:: parts.t2t_timing_signal
+    :members:
+    :undoc-members:
+    :show-inheritance:
+
+utils
+------------------
+
+.. automodule:: parts.utils
+    :members:
+    :undoc-members:
+    :show-inheritance:
