+++ docs/source/api-docs/running.rst
@@ -1,0 +1,31 @@
+running
+=======
+
+.. automodule:: running
+    :members:
+    :undoc-members:
+    :show-inheritance:
+
+funcs
+---------------------
+
+.. automodule:: running.funcs
+    :members:
+    :undoc-members:
+    :show-inheritance:
+
+model\_builders
+-------------------------------
+
+.. automodule:: running.model_builders
+    :members:
+    :undoc-members:
+    :show-inheritance:
+
+utils
+---------------------
+
+.. automodule:: running.utils
+    :members:
+    :undoc-members:
+    :show-inheritance:
