+++ docs/source/conf.py
@@ -20,6 +20,8 @@ import os
 # add these directories to sys.path here. If the directory is relative to the
 # documentation root, use os.path.abspath to make it absolute, like shown here.
 sys.path.insert(0, os.path.abspath('../../'))
+sys.path.insert(0, os.path.abspath('../../open_seq2seq'))
+
 
 # -- General configuration ------------------------------------------------
