+++ run.py
@@ -14,8 +14,8 @@ import copy
 import os
 
 from open_seq2seq.utils import deco_print
-from open_seq2seq.training import train, infer, evaluate
-from open_seq2seq.training.model_builders import \
+from open_seq2seq.running import train, infer, evaluate
+from open_seq2seq.running.model_builders import \
   create_encoder_decoder_loss_model
